separating graphs and charts and rendering the components in home rather than having neverending blocks of code
