+++ src/components/home/Home.js
@@ -4,6 +4,7 @@ import "./Home.css"
 import Table from 'react-bootstrap/Table'
 import KidManager from '../../modules/KidManager'
 import {Line} from 'react-chartjs-2'
+import LineGraph from "../graph/LineGraph"
 
 //TODO:: UTILIZE CHART.JS TO RENDER DATA INTO GRAPHS
 
@@ -11,35 +12,40 @@ class Home extends Component {
 
   state = {
     kids: [],
-    labels: ['January', 'February', 'March',
-    'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'],
-datasets: [
-{
-label: 'Meal Requests',
-fill: false,
-lineTension: 0.5,
-backgroundColor: 'rgba(75,192,192,1)',
-borderColor: 'rgba(0,0,0,1)',
-borderWidth: 2,
-data: []
-}
-]
+    //LINE GRAPH INFO TO TRACK MEAL REQUESTS BY MONTH
+//     labels: ['January', 'February', 'March',
+//     'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'],
+// datasets: [
+// {
+// label: "Meal Requests",
+// fill: false,
+// lineTension: 0.5,
+// backgroundColor: 'rgba(75,192,192,1)',
+// borderColor: 'rgba(0,0,0,1)',
+// borderWidth: 2,
+// data: []
+// }
+// ]
 }
 
 componentDidMount() {
 
   //getAll from KidManager and hang on to that data; put it in state
+  //FOR LOOP HELPS PUSH INFO INTO DATA(STATE) TO PLOT POINTS THAT UPDATE AUTOMAGICALLY IN LINE GRAPH FOR MEAL REQUESTS
   KidManager.getAll()
       .then((kids) => {
-        let data = [];
-        for(let i=0; i<12; i++){
-         const point = kids.filter(kid => new Date(kid.date).getMonth() === i).length
-         data.push(point)
-         console.log(data)
-        }
+        // let data = [];
+        // for(let i=0; i<12; i++){
+        //  const point = kids.filter(kid => new Date(kid.date).getMonth() === i).length
+        //  data.push(point)
+        
+        // }
+        // let copyOfDataSets = this.state.datasets
+        // copyOfDataSets[0].data = data
+        // console.log(copyOfDataSets)
           this.setState({
               kids: kids,
-              datasets:[{"data": data}]
+              // datasets: copyOfDataSets
           
           })
       })
@@ -208,7 +214,7 @@ componentDidMount() {
 </div>
 
 {/* CHART JS ATTEMPTS */}
-<div className="line-graph">
+{/* <div className="line-graph">
         <Line
           data={this.state}
           options={{
@@ -223,7 +229,9 @@ componentDidMount() {
             }
           }}
         />
-      </div>
+      </div> */}
+
+      <LineGraph/>
 
 </>
     )
